repo

+++ src/main/java/com/hcl/sandwich/repository/ItemRepository.java
@@ -2,6 +2,8 @@ package com.hcl.sandwich.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Items, Integer>{
+import com.hcl.sandwich.entity.Items;
+
+public interface ItemRepository extends JpaRepository<Items, Long>{
 
 }
