+++ src/main/java/com/hcl/sandwich/repository/OrderItemRepository.java
@@ -2,6 +2,8 @@ package com.hcl.sandwich.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItems, Integer>{
+import com.hcl.sandwich.entity.OrderItems;
+
+public interface OrderItemRepository extends JpaRepository<OrderItems, Long>{
 
 }
